refactor(loader): split static loader source infrastructure
